+++ CHANGELOG.md
@@ -28,8 +28,11 @@ Classify the change according to the following categories:
 - Add **REoptLogger** type of global logger with a standard out to the console and to a dictionary
     - Instantiate `logREopt` as the global logger in `__init()__` function call as a global variable
     - Handle Warn or Error logs to save them along with information on where they occurred
-    - Add Warnings and Errors from `logREopt` to results dictionary in `core/reopt.jl -> run_reopt()`
-- Add `handle_errors()` to `core/reopt.jl` to gracefully handle scenarios where REopt fails during data processing/optimization
+    - Try-catch `core/reopt.jl -> run_reopt()` functions. Process any errors when catching the error.
+    - Add Warnings and Errors from `logREopt` to results dictionary. If error is unhandled in REopt, include a stacktrace
+    - Add a `status` of `error` to results for consistency
+    - Ensure all error text is returned as strings for proper handling in the API
+- Add `handle_errors(e::E, stacktrace::V) where {E <: Exception, V <: Vector}` and `handle_errors()` to `core/reopt.jl` to gracefully handle scenarios where REopt fails during data processing/optimization
 ### Changed
 - `core/reopt.jl` added try-catch statements to gracefully catch a REopt error and return it to user along with where the error happened.